requirements.txt
django
drf-yasg
djangorestframework-simplejwt
psycopg2-binary
pillow
djangorestframework
